Turning `SearchJob(Store|Actions)` into plain functions.
