+++ graylog2-web-interface/src/views/stores/SearchStore.ts
@@ -23,7 +23,7 @@ import { qualifyUrl } from 'util/URLUtils';
 import fetch from 'logic/rest/FetchProvider';
 import { SearchExecutionStateStore } from 'views/stores/SearchExecutionStateStore';
 import { SearchMetadataActions } from 'views/stores/SearchMetadataStore';
-import { SearchJobActions } from 'views/stores/SearchJobStore';
+import { runSearchJob, searchJobStatus } from 'views/stores/SearchJobs';
 import { ViewStore, ViewActions } from 'views/stores/ViewStore';
 import SearchResult from 'views/logic/SearchResult';
 import SearchActions, { WidgetsToSearch } from 'views/actions/SearchActions';
@@ -145,18 +145,18 @@ export const SearchStore: Store<SearchStoreState> = singletonStore(
         }
 
         return resolve(Bluebird.delay(250)
-          .then(() => SearchJobActions.jobStatus(job.id))
+          .then(() => searchJobStatus(job.id))
           .then((jobStatus) => this.trackJobStatus(jobStatus, search)));
       });
     },
 
     trackJob(search: Search, executionState: SearchExecutionState): Promise<SearchResult> {
-      return SearchJobActions.run(search, executionState).then((job) => this.trackJobStatus(job, search));
+      return runSearchJob(search, executionState).then((job) => this.trackJobStatus(job, search));
     },
 
     execute(executionState: SearchExecutionState): Promise<SearchExecutionResult> {
       const handleSearchResult = (searchResult: SearchResult) => searchResult;
-      const startActionPromise = (executePromise) => SearchActions.execute.promise(executePromise);
+      const startActionPromise = (executePromise: Promise<SearchExecutionResult>) => SearchActions.execute.promise(executePromise);
 
       return this._executePromise(executionState, startActionPromise, handleSearchResult);
     },
@@ -213,7 +213,7 @@ export const SearchStore: Store<SearchStoreState> = singletonStore(
       handleSearchResult: (result: SearchResult) => SearchResult): Promise<SearchExecutionResult> {
       const { widgetsToSearch } = this._state();
 
-      if (this.executePromise && this.executePromise.cancel) {
+      if (this.executePromise?.cancel) {
         this.executePromise.cancel();
       }
